Separate Pydantic AI cookbook example from helper module
