+++ examples/cookbook/pydantic-ai/moss_pydantic_ai.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import asyncio
 import logging
-import os
 from collections.abc import Sequence
 from typing import Any
 
@@ -131,48 +130,3 @@ def as_tool(
         alpha=alpha,
     )
     return moss, moss.tool
-
-
-def _require_env(name: str) -> str:
-    value = os.getenv(name)
-    if not value:
-        raise RuntimeError(f"Missing required environment variable: {name}")
-    return value
-
-
-async def main() -> None:
-    """Run an interactive example using the Moss-backed search tool."""
-    from dotenv import load_dotenv
-    from pydantic_ai import Agent
-
-    load_dotenv()
-
-    client = MossClient(
-        _require_env("MOSS_PROJECT_ID"),
-        _require_env("MOSS_PROJECT_KEY"),
-    )
-    moss = MossSearchTool(
-        client=client,
-        index_name=_require_env("MOSS_INDEX_NAME"),
-    )
-    await moss.load_index()
-
-    agent = Agent(
-        _require_env("PYDANTIC_AI_MODEL"),
-        system_prompt=(
-            "Answer user questions with the help of Moss search when "
-            "factual lookup from the knowledge base is needed."
-        ),
-        tools=[moss.tool],
-    )
-
-    question = os.getenv(
-        "PYDANTIC_AI_PROMPT",
-        "How to reset password?",
-    )
-    result = await agent.run(question)
-    print(result.output)
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
